Fixes sign warning.

+++ samples/look_at/sample_look_at.cc
@@ -254,7 +254,7 @@ class LookAtSampleApplication : public ozz::sample::Application {
     // Showing joints
     if (show_joints_) {
       const float kSphereRadius = .02f;
-      for (size_t i = 0; i < chain_length_; ++i) {
+      for (int i = 0; i < chain_length_; ++i) {
         const ozz::math::Float4x4& transform = models_[joints_chain_[i]];
         success &= _renderer->DrawAxes(transform * kAxesScale);
         success &= _renderer->DrawSphereIm(kSphereRadius, transform,
